requirements.txt
Flask  
datetime 
uuid
Flask-SQLAlchemy
flask_bcrypt
flask_login
PyJWT